+++ src/Farmer/Arm/ContainerService.fs
@@ -62,12 +62,18 @@ type ManagedCluster =
                 yield! this.Identity.Dependencies
             ]
             {| managedClusters.Create(this.Name, this.Location, dependencies) with
-                   identity = this.Identity.ToArmJson
+                   identity = // If using MSI but no identity was set, then enable the system identity like the CLI
+                       if this.ServicePrincipalProfile.ClientId = "msi"
+                          && this.Identity.SystemAssigned = FeatureFlag.Disabled
+                          && this.Identity.UserAssigned.Length = 0 then
+                           { SystemAssigned = Enabled; UserAssigned = [] }.ToArmJson
+                       else
+                           this.Identity.ToArmJson
                    properties =
                        {| agentPoolProfiles =
                            this.AgentPoolProfiles
                            |> List.mapi (fun idx agent ->
-                               {| name = if agent.Name = ResourceName.Empty then $"{this.Name.Value}-agent-pool{idx}"
+                               {| name = if agent.Name = ResourceName.Empty then $"nodepool{idx + 1}"
                                          else agent.Name.Value.ToLowerInvariant ()
                                   count = agent.Count
                                   maxPods = agent.MaxPods |> Option.toNullable